requirements.txt
pandas
numpy
scikit-learn
xgboost
streamlit
joblib
matplotlib
seaborn
scikeras
tensorflow-cpu
